Create README.md
I enjoy problem solving, especially with others, like everyone holding pieces to the puzzle, putting them in place...each contributing to the completed picture. I have an interest in CyberSecurity, as I have been a victim myself. It is a rampant problem which ultimately costs us all, and I am inclined to help reduce the problem and catch the bad guys